+++ FuelPlease/src/main/webapp/WEB-INF/views/caraccount/carAccount.jsp
@@ -1,112 +1,45 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
+	pageEncoding="UTF-8"%>
+
 <!-- CSS -->
-<link rel="stylesheet" href="${pageContext.request.contextPath}/css/carAccount.css">
+<link rel="stylesheet"
+	href="${pageContext.request.contextPath}/css/caraccount/carAccount.css">
 <!-- 폰트어썸 -->
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css"/>
-<!-- 구글폰트 -->
-<link href="https://fonts.googleapis.com/css2?family=Quicksand:wght@300;400;500;600;700&display=swap" rel="stylesheet"/>
+<link rel="stylesheet"
+	href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css" />
 
+<%@ include file="../include/header.jsp"%>
+<section>
+	<div class="calendar-container">
+		<div class="calendar">
+			<div class="month">
+				<i class="fas fa-angle-left prev"></i>
+				<div class="date">
+					<h1></h1>
+					<p></p>
+				</div>
+				<i class="fas fa-angle-right next"></i>
+			</div>
+			<div class="weekdays">
+				<div class="sunday">Sun</div>
+				<div>Mon</div>
+				<div>Tue</div>
+				<div>Wed</div>
+				<div>Thu</div>
+				<div>Fri</div>
+				<div class="saturday">Sat</div>
+			</div>
+			<div class="days"></div>
+		</div>
+	</div>
+</section>
+<%@ include file="../include/footer.jsp"%>
+<script src="${pageContext.request.contextPath}/js/caraccount/carAccount.js"></script>
 
-<body>
-    <div class="container">
-        <div class="calendar">
-            <div class="month">
-                <i class="fas fa-angle-left prev"></i>
-                <div class="date">
-                    <h1></h1>
-                    <p></p>
-                </div>
-                <i class="fas fa-angle-right next"></i>
-            </div>
-            <div class="weekdays">
-                <div>Sun</div>
-                <div>Mon</div>
-                <div>Tue</div>
-                <div>Wed</div>
-                <div>Thu</div>
-                <div>Fri</div>
-                <div>Sat</div>
-            </div>
-            <div class="days">                
-            </div>
-        </div>
-    </div>
-</body>
+<script>
+    document.querySelector('.days').addEventListener('click', (e) => {
+        const value = e.target.textContent;
+        console.log(value);
+    })
 
-<!-- <script>
-    const date = new Date();
-
-    const renderCalendar = () => {
-        date.setDate(1);
-
-        const monthDays = document.querySelector(".days");
-
-        const lastDay = new Date(
-            date.getFullYear(), date.getMonth() + 1, 0).getDate();
-
-        const prevLastDay = new Date(
-            date.getFullYear(), date.getMonth(), 0).getDate();
-
-        const firstDayIndex = date.getDay();
-
-        const lastDayIndex = new Date(
-            date.getFullYear(), date.getMonth() + 1, 0).getDay();
-
-        const nextDays = 7 - lastDayIndex - 1;
-
-        const months = [
-            "January",
-            "February",
-            "March",
-            "April",
-            "May",
-            "June",
-            "July",
-            "August",
-            "September",
-            "October",
-            "November",
-            "December",
-        ];
-
-        document.querySelector(".date h1").innerHTML = months[date.getMonth()];
-
-        document.querySelector(".date p").innerHTML = new Date().toDateString();
-
-        let days = "";
-
-        for (let x = firstDayIndex-1; x >= 0; x--) {
-            days += `<div class="prev-date">${prevLastDay - x + 1}</div>`;
-        }
-
-        for (let i = 1; i <= lastDay; i++) {
-            if (i === new Date().getDate() && date.getMonth() === new Date().getMonth()) {
-                days += `<div class="today">${i}</div>`;
-            } else {
-                days += `<div>${i}</div>`;
-            }
-        }
-
-        for (let j = 1; j <= nextDays; j++) {
-            days += `<div class="next-date">${j}</div>`;
-        }
-            monthDays.innerHTML = days;
-    };
-
-    document.querySelector(".prev").addEventListener("click", () => {
-    date.setMonth(date.getMonth() - 1);
-    renderCalendar();
-    });
-
-    document.querySelector(".next").addEventListener("click", () => {
-    date.setMonth(date.getMonth() + 1);
-    renderCalendar();
-    });
-
-    renderCalendar();
-</script> -->
-    
-
-
-    
+</script>
